refactor: move prettier via this CLI :tada:
